llamaindex/pdfdoc/requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file
# PyMuPDF
python-dotenv
